requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.22.0
requests==2.28.2
mutagen==1.46.0
pyinstaller==6.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
watchfiles==0.19.0
websockets==11.0.3
fusepy==3.0.1
click==8.1.3
Authlib==1.2.0
httpx==0.24.1
pydantic==2.4.2
sqlalchemy==2.0.23
email-validator==2.0.0
